requirements.txt
quizduell-api==1.0
